requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3